+++ sickbeard/db.py
@@ -27,8 +27,8 @@ import threading
 import sickbeard
 
 from sickbeard import logger
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
+from sickrage.helper.exceptions import ex
 
 db_cons = {}
 db_locks = {}
