+++ sickbeard/event_queue.py
@@ -2,7 +2,7 @@ import threading
 import traceback
 from Queue import Queue, Empty
 from sickbeard import logger
-from sickrage.helper.common import ex
+from sickrage.helper.exceptions import ex
 
 
 class Event:
