+++ sickbeard/metadata/mediabrowser.py
@@ -26,9 +26,9 @@ import generic
 
 from sickbeard import logger, helpers
 
-from sickrage.helper.common import ex, dateFormat
+from sickrage.helper.common import dateFormat
 from sickrage.helper.encoding import ek
-from sickrage.helper.exceptions import ShowNotFoundException
+from sickrage.helper.exceptions import ex, ShowNotFoundException
 
 import xml.etree.cElementTree as etree
