+++ sickbeard/metadata/tivo.py
@@ -26,9 +26,8 @@ import sickbeard
 
 from sickbeard import logger, helpers
 from sickbeard.metadata import generic
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
-from sickrage.helper.exceptions import ShowNotFoundException
+from sickrage.helper.exceptions import ex, ShowNotFoundException
 
 
 class TIVOMetadata(generic.GenericMetadata):
