+++ sickbeard/notifiers/growl.py
@@ -20,7 +20,7 @@ import socket
 
 import sickbeard
 from sickbeard import logger, common
-from sickrage.helper.common import ex
+from sickrage.helper.exceptions import ex
 from libgrowl import gntp
