+++ sickbeard/notifiers/pushover.py
@@ -25,7 +25,7 @@ import time
 import sickbeard
 from sickbeard import logger
 from sickbeard.common import notifyStrings, NOTIFY_SNATCH, NOTIFY_DOWNLOAD, NOTIFY_SUBTITLE_DOWNLOAD, NOTIFY_GIT_UPDATE, NOTIFY_GIT_UPDATE_TEXT
-from sickrage.helper.common import ex
+from sickrage.helper.exceptions import ex
 
 API_URL = "https://api.pushover.net/1/messages.json"
