+++ sickbeard/notifiers/pytivo.py
@@ -23,8 +23,8 @@ from urllib import urlencode
 from urllib2 import Request, urlopen, HTTPError
 
 from sickbeard import logger
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
+from sickrage.helper.exceptions import ex
 
 
 class pyTivoNotifier:
