+++ sickbeard/notifiers/synologynotifier.py
@@ -22,8 +22,8 @@ import sickbeard
 
 from sickbeard import logger
 from sickbeard import common
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
+from sickrage.helper.exceptions import ex
 
 
 class synologyNotifier:
