+++ sickbeard/properFinder.py
@@ -33,8 +33,7 @@ from sickbeard import helpers, logger
 from sickbeard import search
 
 from sickbeard.common import DOWNLOADED, SNATCHED, SNATCHED_PROPER, Quality, cpu_presets
-from sickrage.helper.common import ex
-from sickrage.helper.exceptions import AuthException
+from sickrage.helper.exceptions import AuthException, ex
 from sickrage.show.History import History
 
 from name_parser.parser import NameParser, InvalidNameException, InvalidShowException
