+++ sickbeard/rssfeeds.py
@@ -6,8 +6,8 @@ import urllib
 import sickbeard
 
 from sickbeard import logger
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
+from sickrage.helper.exceptions import ex
 
 from feedcache.cache import Cache
