+++ tests/encoding_tests.py
@@ -9,8 +9,8 @@ sys.path.insert(1, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 
 import sickbeard
 from sickbeard.helpers import sanitizeFileName
-from sickrage.helper.common import ex
 from sickrage.helper.encoding import ek
+from sickrage.helper.exceptions import ex
 
 
 class EncodingTests(unittest.TestCase):
