+++ tests/ssl_sni_tests.py
@@ -25,7 +25,7 @@ sys.path.insert(1, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 import requests
 import sickbeard.providers as providers
 import certifi
-from sickrage.helper.common import ex
+from sickrage.helper.exceptions import ex
 
 
 class SNI_Tests(unittest.TestCase):
